Implement new OpusRouter and add more libopus function bindings
Added set_gain and get_last_packet_duration decoder CTLs and functions.
Add new OpusRouter.  This one is a bit different than the previous one.  It runs as a thread preventing user code from blocking the socket read loop.  This allows it to write packets to sinks in realtime instead of in chunks, allowing for less buffering on the user's end.  The performance hit is probably higher too but that's just how voice recv is going to be unfortunately.  I may consider adding a low cost router later if demand... exists.